Remove redundant code and fix utils & types dirs
