+++ src/features/berth/berthFormPage/switchApplication/switchApplication.scss
