+++ src/features/berth/berthFormPage/switchApplication/types.ts
@@ -1,0 +1,8 @@
+import { BerthSwitchReasonsQuery_berthSwitchReasons } from '../../../__generated__/BerthSwitchReasonsQuery';
+
+export type Option = {
+  label: string;
+  value: string;
+};
+
+export type Reason = BerthSwitchReasonsQuery_berthSwitchReasons;
